+++ app/Http/Middleware/VerifyCsrfToken.php
@@ -6,5 +6,5 @@ use Illuminate\Foundation\Http\Middleware\VerifyCsrfToken as BaseVerifier;
 
 class VerifyCsrfToken extends BaseVerifier
 {
-    protected $except = ['api/login'];
+    protected $except = ['api/login', 'api/password/email', 'api/password/reset'];
 }