+++ resources/assets/js/pages/administration/users/Show.vue
@@ -329,7 +329,6 @@ export default {
         logout() {
             axios.post(route('logout', [], false).toString()).then(() => {
                 this.$store.commit('auth/logout');
-                this.$router.push({ path: '/' });
             }).catch(error => this.handleError(error));
         },
         getDay(timestamp) {