+++ resources/assets/js/pages/layout/Auth.vue
@@ -1,9 +1,9 @@
 <template>
 
-    <section class="hero login is-fullheight is-primary is-bold">
+    <section class="hero is-fullheight is-primary is-bold">
         <div class="hero-body">
             <div class="container">
-                <div class="columns is-mobile is-centered">
+                <div class="columns is-centered">
                     <router></router>
                 </div>
             </div>